ftrace: Add 'recover' trigger action
The 'recover' trigger action is to restore hooked return address for
some functions.  For example, node.js (hence v8) traces the addresses
when it collects garbages and gets segfaults if it accesses the
pointer.  In this case, users can set 'recover' trigger for the garbage
collector functions so that it can see the correct addresses.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>