uftrace: Change opts->kernel as bool
It's more reasonal to treat this option as a boolean.

But it affects the timeout of writer threads.  I think it's ok as
there's no observable difference in the output - it still has LOST
records before and after the change for some large kernel depth.  I
still need to increse the kernel buffer size to avoid that.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>